object cache evaluation: latex output tables, automatically run 4 configurations (sram,sdram,sram cmp8,sdram cmp8)

+++ java/tools/src/com/jopdesign/wcet/analysis/cache/ObjectCacheEvaluation.java
@@ -49,7 +49,7 @@ public class ObjectCacheEvaluation {
 		 * @param bestCyclesPerAccessForConfig
 		 */
 		public OCacheAnalysisResult(OCacheMode mode, int ways, int lineSize,
-				                    int configId, long hitRate,
+				                    int configId, double hitRate,
 				                    double cyclesPerAccess) {
 			this.mode = mode;
 			this.ways = ways;
